+++ CHANGELOG.md
@@ -49,6 +49,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - FIX: Added `AuthenticationGroups` to the manifest.yaml under samples
 - FIX: cleanup of istio-ingress raises exception when no cluster exists
 - FIX: replace cert-manager/jobs with cert-manager/certifactes
+- FIX: remove the ownerReference on userspace PodDefaults (because they're in the teamspace namespace)
 
 ### **Removed**